+++ cmd2/__init__.py
@@ -10,7 +10,7 @@ except DistributionNotFound:
     # package is not installed
     pass
 
-from .ansi import style
+from .ansi import style, fg, bg
 from .argparse_custom import Cmd2ArgumentParser, CompletionError, CompletionItem, set_default_argument_parser
 
 # Check if user has defined a module that sets a custom value for argparse_custom.DEFAULT_ARGUMENT_PARSER